Update jest-snapshot to version 16
This update of jest changes the snapshot state api. Its not jasmine specific now. So we can get rid of faking jasmine. It takes away the snapshot.add() method and does most of the snapshot processing inside the state.match() method.

This makes it a little difficult to differentiate states between ‘added’, ‘updated’ and ‘matched’. Added a method `getTestOutcome` to do that.