Remove broken react-button example
The [react-button](http://kadira-samples.github.io/react-button/) storybook is linked to as an example here and on the homepage, but it [appears very broken now](https://github.com/kadira-samples/react-button/issues/15). As the first example I opened, this totally confused me as I wasn't sure if what I was seeing was intended. I think this should be removed until fixed so people don't get a bad first impression like I did, and the other examples are already pretty good.