Set up test-cra app to use storyshots
A few things involved:

1. Remove dependencies from storyshots

Storyshots wants to plugin to the version used by the app, not have its own version. So it should probably use a peer dep; however as some deps are optional (i.e. react vs react-native) this will lead to confusing warnings to the user.

This seems like an OK approach until we hear differently from users.

2. Update test-cra's react-scripts to get it's Jest version up to date

3. Add a `mode` option to storyshots to avoid it auto-detecting if `@storybooks/react[-native]` is installed (this was breaking Jest run from the top-level).

4. Ensure that travis installs the `test-cra` properly