Do not ignore story.js in prepublish
This is causing an issue with addons-info as it ends up ignoring `Story.js` during compilation. Seems there are case-similar files in other addon packages called `story.js` that were supposed to be ignored from output... However, this seems like the quickest fix that doesn't really have an impact on other packages.