Add small design update to addon info package
When using the "info" addon inline, the current designs blends in
too much with the component, making it quite confusing and sometimes
difficult to preview.
This change wraps the Story info into a "box" with a light border and
shadow, also adds some more padding and margin to give a bit more room.