Remove `@storybook/react` dependencies from storyshots
This is a hack. Currently the build is broken because `test-cra`
requires `@storybook/addon-storyshots` AND `@storybook/react`, but
`@storybook/addon-storyshots` requires `@storybook/react`. Before this
change, two versions of `@storybook/react` are getting loaded, and this
leads to a test error “Your test suite must contain at least one test.”

After this “fix” the error goes away. However,
`@storybook/addon-storyshots` should have a dev- dependency on
`@storybook/react`, because it uses it for its examples. But it only
has a peer-dep for now.