Emit a custom 'click' event
The story listens for a custom 'click' event emited from `MyButton.vue` component, but `MyButton.vue` lacks such a functionality. When the button is clicked, only an empty method is called.

The change I suggested is the less intrusive one, keeping the method code as opposite to removing it and emitting an event directly from the template.