Add 'waitUntil' option to puppeteer of storyshots
If some elements use additional xhr requests (like async image loading).
Puppeteer might not imprint the end state of the page. Puppeteer has
built-in method to resolve this -
https://github.com/GoogleChrome/puppeteer/blob/c857aef5893042ac2f8b587283683396ae9c333b/docs/api.md#pagegotourl-options

This option is fairly general so it should be okay to use for other cases
as well.

However, 'beforeScreenshot' could be used instead, but proposed approach seems
more definite and reliable

As additional further improvement, these `goto` options might be exposed to
consumber as well