fix: removed dead link
The first of the three links for example storybooks isn't working anymore. The linked storybook just prints errors and is not showing any components. A GitHub issue exists for this (https://github.com/kadira-samples/react-button/issues/17), but the author seems to be irresponsive. Broken example isn't helping anyone and should be removed.