Fix knobs for React < 16.3
The `getDerivedStateFromProps` lifecycle method was added in React 16.3,
and older versions of react won’t execute it.  Many of the knob components
were relying on this function to initialize component state, and so in
older versions of react state was undefined, and grabbing `value` would
blow up.

Most of the usages in these components were to memoize props, but that
can be handled more cleanly through the use of `PureComponent`, as 
mentioned in:
https://reactjs.org/blog/2018/06/07/you-probably-dont-need-derived-state.html#what-about-memoization