Use shouldComponentUpdate instead of PureComponent
Because the `knob` object prop is being regenerated each time the
parent re-renders, we can’t use the shallow prop comparison that 
PureComponent does.  So this uses a `shouldComponentUpdate` instead,
which in most cases just compares the `knob.value` prop to see if it
has changed.  I don’t believe that the `name` is changed between
renders.  Same goes for number range configurations and array
separators.