Fix for issue #4066 / #4180
The webpack currently uses both file-loader and svg-url-loader.

file-loader will handle .svg files just fine. svg-url-loader is causing images to not load properly. This might be identified by @cainlevy in issue 4180:
> svg-url-loader is not drop-in compatible with url-loader.

Unless there is a compelling reason to use svg-url-loader instead of just leveraging the currently utilized url-loader, this change seems to fix the issues and can result in removing the dependency on svg-url-loader from the repo completely.