REFACTOR no scroller around Tabs content, instead force scroll on child
This I hacky, I know but we need to scroll over content, without the Actionbar component, that should stick.

I'd like to refactor the Tabs component so it can render ActionBar independently, but that a bigger overhaul,
then I feel comfortable with at the moment.

Most of our addons will show a custom scrollbar now, but it's a manual process, the addon rendering is taking control
over scrolling.

Instead I'd like tabs to be in control of scrolling, but that's currently not feasable with how ActionBar works.