chore: remove `node-weak` dependency
`node-weak` is not working with Node 12 and so make `yarn bootstrap` command fail.
Moreover it is used only for jest related stuff and only affects some developers.
The good news is Jest has been working on that and the next release of Jest will certainly get ride of this dependency.
See https://github.com/facebook/jest/pull/8686