perf: Apply state colocation principles and avoid prop drilling
In particular the story context state is maintained near the root of
the tree, any changes to this state end up re-rendering a large number
of components, such as the story list. By reading the state closer to
where it needs to be used, it is possible to avoid prop
drilling-inflicted rerenders.

I opted for Jotai here, since its a relatively small and well tested
and maintained state management library, that imposes basically
nothing on an application's existing structure.