fix(ondevice-controls): Arg state can become desynced from the story
This typically happens when the `ControlsPanel` component is updated,
instead of remounted, and `useArgs` is called with `initialArgs` for
the new story ID but because the `useState` initializer is only run
once, the arg state is not resynced when the story changes.

This is now explicitly done when the story ID changes.