fix: StoryView was interfering with scroll gestures
This fix works by using the React Native responder system to detect
the interaction, and dismiss the keyboard, but declines to actually
start handling the gesture. This means that the attempt is handed
to the next component in the hierarchy to handle.

Unlike `Touchable...` which _wants_ to handle the interaction, and
ends up interfering with interactions.